Safari-611.1.10/LayoutTests/http/tests/referrer-policy-script/strict-origin-when-cross-origin/same-origin.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of strict-origin-when-cross-origin referrer policy when same origin.");
jsTestIsAsync = true;

function checkReferrer(value) {
    referrer = value;
    // Should be the full URL because we are same-origin.
    shouldBeEqualToString("referrer", "http://127.0.0.1:8000/referrer-policy-script/strict-origin-when-cross-origin/same-origin.html");
    finishJSTest();
}

if (window.testRunner) {
    testRunner.setStatisticsShouldDowngradeReferrer(false, function () { downgradeReferrerCallback('strict-origin-when-cross-origin', '127.0.0.1')});
}
</script>
</body>
</html>
